requirements.txt
streamlit
tensorflow-macos
tensorflow-metal
pillow
numpy
pandas
matplotlib
